# Conflicts:
#	gateway/src/main/java/ru/practicum/shareit/item/api/controller/ItemController.java
#	gateway/src/main/resources/application.properties
#	server/src/main/resources/application.properties

+++ gateway/src/main/java/ru/practicum/shareit/booking/api/controller/BookingController.java
@@ -39,7 +39,7 @@ public class BookingController {
 
     @PostMapping(CREATE_BOOKING)
     public ResponseEntity<Object> create(
-            @RequestHeader(HEADER_USER_ID) Integer userId,
+            @RequestHeader(HEADER_USER_ID) Long userId,
             @Valid @RequestBody BookingSimpleDto bookingSimpleDto) {
 
         return bookingClient.create(userId, bookingSimpleDto);
@@ -47,7 +47,7 @@ public class BookingController {
 
     @PatchMapping(UPDATE_STATUS_BOOKING)
     public ResponseEntity<Object> update(
-            @RequestHeader(HEADER_USER_ID) Integer userId,
+            @RequestHeader(HEADER_USER_ID) Long userId,
             @PathVariable Long bookingId,
             @RequestParam boolean approved) {
 
@@ -56,7 +56,7 @@ public class BookingController {
 
     @GetMapping(GET_BOOKING)
     public ResponseEntity<Object> get(
-            @RequestHeader(HEADER_USER_ID) Integer userId,
+            @RequestHeader(HEADER_USER_ID) Long userId,
             @PathVariable Long id) {
 
         return bookingClient.getById(userId, id);
@@ -64,7 +64,7 @@ public class BookingController {
 
     @GetMapping(GET_ALL_BOOKINGS_FOR_USER)
     public ResponseEntity<Object> getAllByUser(
-            @RequestHeader(HEADER_USER_ID) Integer userId,
+            @RequestHeader(HEADER_USER_ID) Long userId,
             @RequestParam(required = false, defaultValue = "ALL") String stateIn,
             @RequestParam(required = false, defaultValue = FROM)
             @PositiveOrZero Integer from,
@@ -78,7 +78,7 @@ public class BookingController {
 
     @GetMapping(GET_ALL_BOOKINGS_FOR_OWNER)
     public ResponseEntity<Object> getAllByOwner(
-            @RequestHeader(HEADER_USER_ID) Integer userId,
+            @RequestHeader(HEADER_USER_ID) Long userId,
             @RequestParam(required = false, defaultValue = "ALL") BookingState state,
             @RequestParam(required = false, defaultValue = FROM)
             @PositiveOrZero Integer from,
